backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Validation & Settings
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1

# Firebase Admin
firebase-admin==6.6.0

# Google Gemini AI
google-genai>=1.0.0

# Stripe
stripe==11.0.0

# PDF & Document Processing
pypdf2==3.0.1
python-docx==1.1.0
reportlab>=4.0.0

# Image Processing
Pillow>=10.0.0

# HTTP Client
httpx==0.27.0

# Web Scraping (job URLs)
beautifulsoup4==4.12.3

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Development
black==24.10.0
isort==5.13.2
